refactor logging so it is same in all scripts
