--- hack/log/log-dump-daemonset-windows.yaml
@@ -1,36 +1,0 @@
-apiVersion: apps/v1
-kind: DaemonSet
-metadata:
-  name: log-dump-node-windows
-spec:
-  selector:
-    matchLabels:
-      app: log-dump-node-windows
-  template:
-    metadata:
-      labels:
-        app:
-          log-dump-node-windows
-    spec:
-      securityContext:
-        windowsOptions:
-          runAsUserName: ContainerAdministrator
-      containers:
-      - name: log-dump-node-windows
-        image: mcr.microsoft.com/oss/kubernetes/pause:3.6
-        volumeMounts:
-        - name: varlog
-          mountPath: /var/log
-      nodeSelector:
-        kubernetes.io/os: windows
-      tolerations:
-      - effect: NoExecute
-        operator: Exists
-      - effect: NoSchedule
-        operator: Exists
-      - key: CriticalAddonsOnly
-        operator: Exists
-      volumes:
-      - name: varlog
-        hostPath:
-          path: /var/log
