+++ test/e2e/e2e_suite_test.go
@@ -20,229 +20,23 @@ limitations under the License.
 package e2e
 
 import (
-	"bufio"
 	"context"
-	"encoding/json"
 	"flag"
 	"fmt"
-	"io"
 	"os"
-	"path"
 	"path/filepath"
 	"strings"
 	"testing"
-	"time"
 
-	aadpodv1 "github.com/Azure/aad-pod-identity/pkg/apis/aadpodidentity/v1"
-	"github.com/Azure/azure-sdk-for-go/services/preview/monitor/mgmt/2019-06-01/insights"
-	"github.com/Azure/go-autorest/autorest/azure/auth"
-	"github.com/Azure/go-autorest/autorest/to"
 	. "github.com/onsi/ginkgo"
 	"github.com/onsi/ginkgo/config"
 	"github.com/onsi/ginkgo/reporters"
 	. "github.com/onsi/gomega"
-	corev1 "k8s.io/api/core/v1"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/runtime"
-	"k8s.io/apimachinery/pkg/runtime/schema"
-	infrav1 "sigs.k8s.io/cluster-api-provider-azure/api/v1beta1"
-	expv1 "sigs.k8s.io/cluster-api-provider-azure/exp/api/v1beta1"
-	clusterv1exp "sigs.k8s.io/cluster-api/exp/api/v1beta1"
 	capi_e2e "sigs.k8s.io/cluster-api/test/e2e"
 	"sigs.k8s.io/cluster-api/test/framework"
 	"sigs.k8s.io/cluster-api/test/framework/bootstrap"
 	"sigs.k8s.io/cluster-api/test/framework/clusterctl"
-	"sigs.k8s.io/controller-runtime/pkg/client"
 )
-
-type (
-	AzureClusterProxy struct {
-		framework.ClusterProxy
-	}
-	// myEventData is used to be able to Marshal insights.EventData into JSON
-	// see https://github.com/Azure/azure-sdk-for-go/issues/8224#issuecomment-614777550
-	myEventData insights.EventData
-)
-
-func NewAzureClusterProxy(name string, kubeconfigPath string, scheme *runtime.Scheme, options ...framework.Option) *AzureClusterProxy {
-	proxy, ok := framework.NewClusterProxy(name, kubeconfigPath, scheme, options...).(framework.ClusterProxy)
-	Expect(ok).To(BeTrue(), "framework.NewClusterProxy must implement capi_e2e.ClusterProxy")
-	return &AzureClusterProxy{
-		ClusterProxy: proxy,
-	}
-}
-
-func (acp *AzureClusterProxy) CollectWorkloadClusterLogs(ctx context.Context, namespace, name, outputPath string) {
-	Byf("Dumping workload cluster %s/%s logs", namespace, name)
-	acp.ClusterProxy.CollectWorkloadClusterLogs(ctx, namespace, name, outputPath)
-
-	aboveMachinesPath := strings.Replace(outputPath, "/machines", "", 1)
-
-	Byf("Dumping workload cluster %s/%s kube-system pod logs", namespace, name)
-	start := time.Now()
-	acp.collectPodLogs(ctx, namespace, name, aboveMachinesPath)
-	Byf("Fetching kube-system pod logs took %s", time.Since(start).String())
-
-	Byf("Dumping workload cluster %s/%s Azure activity log", namespace, name)
-	start = time.Now()
-	acp.collectActivityLogs(ctx, aboveMachinesPath)
-	Byf("Fetching activity logs took %s", time.Since(start).String())
-}
-
-func (acp *AzureClusterProxy) collectPodLogs(ctx context.Context, namespace string, name string, aboveMachinesPath string) {
-	workload := acp.GetWorkloadCluster(ctx, namespace, name)
-	pods := &corev1.PodList{}
-	Expect(workload.GetClient().List(ctx, pods, client.InNamespace(kubesystem))).To(Succeed())
-
-	events, err := workload.GetClientSet().CoreV1().Events(kubesystem).List(ctx, metav1.ListOptions{})
-	eventMsgs := map[string]string{}
-	if err != nil {
-		Byf("failed to get events in kube-system namespace: %v", err)
-	} else {
-		for _, event := range events.Items {
-			if event.InvolvedObject.Kind == "Pod" {
-				if _, ok := eventMsgs[event.InvolvedObject.Name]; !ok {
-					eventMsgs[event.InvolvedObject.Name] = event.Message
-				} else {
-					eventMsgs[event.InvolvedObject.Name] += fmt.Sprintf("\n%s", event.Message)
-				}
-			}
-		}
-	}
-
-	for _, pod := range pods.Items {
-		for _, container := range pod.Spec.Containers {
-			// Watch each container's logs in a goroutine so we can stream them all concurrently.
-			go func(pod corev1.Pod, container corev1.Container) {
-				defer GinkgoRecover()
-
-				Byf("Creating log watcher for controller %s/%s, container %s", kubesystem, pod.Name, container.Name)
-				logFile := path.Join(aboveMachinesPath, kubesystem, pod.Name, container.Name+".log")
-				if os.MkdirAll(filepath.Dir(logFile), 0755); err != nil {
-					// Failing to mkdir should not cause the test to fail
-					Byf("Error mkdir: %v", err)
-					return
-				}
-
-				f, err := os.OpenFile(logFile, os.O_APPEND|os.O_CREATE|os.O_WRONLY, 0644)
-				if err != nil {
-					// Failing to fetch logs should not cause the test to fail
-					Byf("Error opening file to write pod logs: %v", err)
-					return
-				}
-				defer f.Close()
-
-				opts := &corev1.PodLogOptions{
-					Container: container.Name,
-					Follow:    true,
-				}
-
-				podLogs, err := workload.GetClientSet().CoreV1().Pods(kubesystem).GetLogs(pod.Name, opts).Stream(ctx)
-				if err != nil {
-					// Failing to stream logs should not cause the test to fail
-					Byf("Error starting logs stream for pod %s/%s, container %s: %v", kubesystem, pod.Name, container.Name, err)
-					return
-				}
-				defer podLogs.Close()
-
-				out := bufio.NewWriter(f)
-				defer out.Flush()
-				_, err = out.ReadFrom(podLogs)
-				if err != nil && err != io.ErrUnexpectedEOF {
-					// Failing to stream logs should not cause the test to fail
-					Byf("Got error while streaming logs for pod %s/%s, container %s: %v", kubesystem, pod.Name, container.Name, err)
-				}
-			}(pod, container)
-		}
-
-		go func(pod corev1.Pod) {
-			defer GinkgoRecover()
-
-			Byf("Collecting events for Pod %s/%s", kubesystem, pod.Name)
-			eventFile := path.Join(aboveMachinesPath, kubesystem, pod.Name, "pod-events.txt")
-			if err := os.MkdirAll(filepath.Dir(eventFile), 0755); err != nil {
-				// Failing to mkdir should not cause the test to fail
-				Byf("Error mkdir: %v", err)
-				return
-			}
-
-			f, err := os.OpenFile(eventFile, os.O_APPEND|os.O_CREATE|os.O_WRONLY, 0644)
-			if err != nil {
-				// Failing to open the file should not cause the test to fail
-				Byf("Error opening file to write Pod events: %v", err)
-				return
-			}
-			defer f.Close()
-
-			msg, ok := eventMsgs[pod.Name]
-			if !ok {
-				Byf("failed to find events of Pod %q", pod.Name)
-				return
-			}
-
-			out := bufio.NewWriter(f)
-			defer out.Flush()
-			_, err = out.WriteString(msg)
-			if err != nil && err != io.ErrUnexpectedEOF {
-				// Failing to collect event message should not cause the test to fail
-				Byf("failed to collect event message of pod %s/%s: %v", kubesystem, pod.Name, err)
-			}
-		}(pod)
-	}
-}
-
-func (acp *AzureClusterProxy) collectActivityLogs(ctx context.Context, aboveMachinesPath string) {
-	timeoutctx, cancel := context.WithTimeout(ctx, 30*time.Second)
-	defer cancel()
-	settings, err := auth.GetSettingsFromEnvironment()
-	Expect(err).NotTo(HaveOccurred())
-	subscriptionID := settings.GetSubscriptionID()
-	authorizer, err := settings.GetAuthorizer()
-	Expect(err).NotTo(HaveOccurred())
-	activityLogsClient := insights.NewActivityLogsClient(subscriptionID)
-	activityLogsClient.Authorizer = authorizer
-
-	groupName := os.Getenv(AzureResourceGroup)
-	start := time.Now().Add(-2 * time.Hour).UTC().Format(time.RFC3339)
-	end := time.Now().UTC().Format(time.RFC3339)
-
-	itr, err := activityLogsClient.ListComplete(timeoutctx, fmt.Sprintf("eventTimestamp ge '%s' and eventTimestamp le '%s' and resourceGroupName eq '%s'", start, end, groupName), "")
-	if err != nil {
-		// Failing to fetch logs should not cause the test to fail
-		Byf("Error fetching activity logs for resource group %s: %v", groupName, err)
-		return
-	}
-
-	logFile := path.Join(aboveMachinesPath, activitylog, groupName+".log")
-	Expect(os.MkdirAll(filepath.Dir(logFile), 0755)).To(Succeed())
-
-	f, err := os.OpenFile(logFile, os.O_APPEND|os.O_CREATE|os.O_WRONLY, 0644)
-	if err != nil {
-		// Failing to fetch logs should not cause the test to fail
-		Byf("Error opening file to write activity logs: %v", err)
-		return
-	}
-	defer f.Close()
-	out := bufio.NewWriter(f)
-	defer out.Flush()
-
-	for ; itr.NotDone(); err = itr.NextWithContext(timeoutctx) {
-		if err != nil {
-			Byf("Got error while iterating over activity logs for resource group %s: %v", groupName, err)
-			return
-		}
-		event := itr.Value()
-		if to.String(event.Category.Value) != "Policy" {
-			b, err := json.MarshalIndent(myEventData(event), "", "    ")
-			if err != nil {
-				Byf("Got error converting activity logs data to json: %v", err)
-			}
-			if _, err = out.WriteString(string(b) + "\n"); err != nil {
-				Byf("Got error while writing activity logs for resource group %s: %v", groupName, err)
-			}
-		}
-	}
-}
 
 func init() {
 	flag.StringVar(&configPath, "e2e.config", "", "path to the e2e config file")
@@ -270,9 +64,6 @@ var _ = SynchronizedBeforeSuite(func() []byte {
 	Expect(configPath).To(BeAnExistingFile(), "Invalid test suite argument. e2e.config should be an existing file.")
 	Expect(os.MkdirAll(artifactFolder, 0755)).To(Succeed(), "Invalid test suite argument. Can't create e2e.artifacts-folder %q", artifactFolder)
 
-	By("Initializing a runtime.Scheme with all the GVK relevant for this test")
-	scheme := initScheme()
-
 	Byf("Loading the e2e test configuration from %q", configPath)
 	e2eConfig = loadE2EConfig(configPath)
 
@@ -280,7 +71,7 @@ var _ = SynchronizedBeforeSuite(func() []byte {
 	clusterctlConfigPath = createClusterctlLocalRepository(e2eConfig, filepath.Join(artifactFolder, "repository"))
 
 	By("Setting up the bootstrap cluster")
-	bootstrapClusterProvider, bootstrapClusterProxy = setupBootstrapCluster(e2eConfig, scheme, useExistingCluster)
+	bootstrapClusterProvider, bootstrapClusterProxy = setupBootstrapCluster(e2eConfig, useExistingCluster)
 
 	By("Initializing the bootstrap cluster")
 	initBootstrapCluster(bootstrapClusterProxy, e2eConfig, clusterctlConfigPath, artifactFolder)
@@ -305,8 +96,7 @@ var _ = SynchronizedBeforeSuite(func() []byte {
 	kubeconfigPath := parts[3]
 
 	e2eConfig = loadE2EConfig(configPath)
-	bootstrapClusterProxy = NewAzureClusterProxy("bootstrap", kubeconfigPath, initScheme(),
-		framework.WithMachineLogCollector(AzureLogCollector{}))
+	bootstrapClusterProxy = NewAzureClusterProxy("bootstrap", kubeconfigPath, framework.WithMachineLogCollector(AzureLogCollector{}))
 })
 
 // Using a SynchronizedAfterSuite for controlling how to delete resources shared across ParallelNodes (~ginkgo threads).
@@ -322,28 +112,6 @@ var _ = SynchronizedAfterSuite(func() {
 		tearDown(bootstrapClusterProvider, bootstrapClusterProxy)
 	}
 })
-
-func initScheme() *runtime.Scheme {
-	scheme := runtime.NewScheme()
-	framework.TryAddDefaultSchemes(scheme)
-	Expect(infrav1.AddToScheme(scheme)).To(Succeed())
-	Expect(expv1.AddToScheme(scheme)).To(Succeed())
-	Expect(clusterv1exp.AddToScheme(scheme)).To(Succeed())
-	// Add aadpodidentity v1 to the scheme.
-	aadPodIdentityGroupVersion := schema.GroupVersion{Group: aadpodv1.GroupName, Version: "v1"}
-	scheme.AddKnownTypes(aadPodIdentityGroupVersion,
-		&aadpodv1.AzureIdentity{},
-		&aadpodv1.AzureIdentityList{},
-		&aadpodv1.AzureIdentityBinding{},
-		&aadpodv1.AzureIdentityBindingList{},
-		&aadpodv1.AzureAssignedIdentity{},
-		&aadpodv1.AzureAssignedIdentityList{},
-		&aadpodv1.AzurePodIdentityException{},
-		&aadpodv1.AzurePodIdentityExceptionList{},
-	)
-	metav1.AddToGroupVersion(scheme, aadPodIdentityGroupVersion)
-	return scheme
-}
 
 func loadE2EConfig(configPath string) *clusterctl.E2EConfig {
 	config := clusterctl.LoadE2EConfig(context.TODO(), clusterctl.LoadE2EConfigInput{ConfigPath: configPath})
@@ -371,7 +139,7 @@ func createClusterctlLocalRepository(config *clusterctl.E2EConfig, repositoryFol
 	return clusterctlConfig
 }
 
-func setupBootstrapCluster(config *clusterctl.E2EConfig, scheme *runtime.Scheme, useExistingCluster bool) (bootstrap.ClusterProvider, framework.ClusterProxy) {
+func setupBootstrapCluster(config *clusterctl.E2EConfig, useExistingCluster bool) (bootstrap.ClusterProvider, framework.ClusterProxy) {
 	var clusterProvider bootstrap.ClusterProvider
 	kubeconfigPath := ""
 	if !useExistingCluster {
@@ -386,8 +154,8 @@ func setupBootstrapCluster(config *clusterctl.E2EConfig, scheme *runtime.Scheme,
 		Expect(kubeconfigPath).To(BeAnExistingFile(), "Failed to get the kubeconfig file for the bootstrap cluster")
 	}
 
-	clusterProxy := NewAzureClusterProxy("bootstrap", kubeconfigPath, scheme)
-	Expect(clusterProxy).ToNot(BeNil(), "Failed to get a bootstrap cluster proxy")
+	clusterProxy := NewAzureClusterProxy("bootstrap", kubeconfigPath)
+	Expect(clusterProxy).NotTo(BeNil(), "Failed to get a bootstrap cluster proxy")
 
 	return clusterProvider, clusterProxy
 }
